HUG/gemini-image-editing/requirements.txt
gradio
google-genai
Pillow
asyncio
pydantic==2.10.6